master pusher ruberger graudolf

+++ src/main/java/client/montecarlo/IMoveController.java
@@ -4,9 +4,9 @@ import client.montecarlo.ActionException;
 import client.montecarlo.SensorDataSet;
 
 public interface IMoveController {
-    void moveForward ( int cm ) throws ActionException;
-    void moveBackward ( int cm)throws ActionException;
-    void turnLeft ( int angle )throws ActionException;
-    void turnRight ( int angle )throws ActionException;
+    void moveForward ( double cm ) throws ActionException;
+    void moveBackward ( double cm)throws ActionException;
+    void turnLeft ( double angle )throws ActionException;
+    void turnRight ( double angle )throws ActionException;
     SensorDataSet getSensorDataSet() throws ActionException;
 }
